% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GibbsGMM}
\alias{GibbsGMM}
\title{wrapper function for the gibbs sampler for the GMM}
\usage{
GibbsGMM(yS, prS, itS, biS, salitS)
}
\arguments{
\item{prS}{is a list of priors for the model}

\item{itS}{max number of iterations}

\item{biS}{number of burn-in iterations}

\item{salitS}{number of allocactions to save in a salso compatible manner (matrix each row is a iteration)}

\item{ys}{the data vector (1D) or matrix (mD)}
}
\value{
List with the posterior samples of each random variable
}
\description{
this identifies whether we are in the 1D or mD case and calls the appropriate function
}
\details{
the prior list prS must include: mu: the prior mean of the normal distr of the mean of each cluster (vector)   mu.sd: the prior sd of the normal distr of the mean of each cluster (vectorsigma the shape parameter of the gamma for the sd prior (vector)sigma.nu the rate parameters of the gamma for the sd prior (vector)
itS total number of iterations
}
