% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MCMCPotts}
\alias{MCMCPotts}
\title{wrapper function for the gibbs sampler for the potts model with beta approximatted by pseudolikelihood}
\usage{
MCMCPotts(yS, nS, bS, itS, biS, prS, mhS, salitS)
}
\arguments{
\item{yS}{a matrix of data, samples on the cols, features on the rows}

\item{nS}{is the neighbours object from BayesImageS}

\item{bS}{is the blocks object from BayesImageS}

\item{itS}{number of iterations to perform}

\item{biS}{number of burn in iterations}

\item{prS}{is a list of priors for the model, it includes:}

\item{mhS}{is a list of options for the MH step when sampling beta}

\item{salitS}{number of allocactions to save in a salso compatible manner (matrix each row is a iteration)}

\item{mu:}{the prior mean of the normal distr of the mean of each cluster (matrix)}

\item{mu.sigma}{the covariance matrices of the normal distr of the mean of each cluster (3D array)}

\item{sigma.V0}{the V= matrix for the wishart prior of the covariance matrices (3D array)}

\item{sigma.n0}{the n0 parameters of the wishart prior for the covariance matrices (vector)}

\item{beta}{constaining the borders of the interval where the beta is uniformly distributed (vector)}
}
\value{
a List with the samples of theposterior of the random variables
}
\description{
this identifies wheter we are in the 1d or md case and calls the appropriate function
}
\details{
mhS it includes:
   bandwith the standard deviation of the normal distr used for the rwmh when updating beta
   init optional value used to initalize the beta to start from
   adaptive optional,if omitted the default is adaptive, if adaptive = NULL then the bandwidth is fixed
   target optional, used to dedtermine how fast to change the bandwidth int the adaptive case (don't change if it is not strictly necessary)
}
