\name{multiPotts-package}
\alias{multiPotts-package}
\alias{multiPotts}
\docType{package}
\title{\packageTitle{multiPotts}}
\description{\packageDescription{multiPotts}}
\section{Package Content}{\packageIndices{multiPotts}}
\author{\packageAuthor{multiPotts}}
\section{Maintainer}{\packageMaintainer{multiPotts}}
\keyword{package}
